+++ modules/client/CHANGELOG.md
@@ -17,6 +17,7 @@ TEMPLATE:
 - Fixed plugin version on the dao subgraph queries
 ### Added
 - Add `getPlugins` function
+- Add `getPlugin` function
 ## [1.5.0]
 ### Added
 - Add `wrapTokens` and `unwrapTokens` functions to support token contract without support for snapshots